+++ src/main/java/com/vik/covid19vik/USTimeSeries.java
@@ -70,7 +70,7 @@ class USTimeSeries {
             private String combinedKey;
             private float lat;
             private float lon;
-            private int uid;
+            private float uid;
             private String iso2;
             private String iso3;
             private int code3;
@@ -112,10 +112,10 @@ class USTimeSeries {
             void setLon(float lon) {
                 this.lon = lon;
             }
-            protected int getUid() {
+            protected float getUid() {
                 return uid;
             }
-            protected void setUid(int uid) {
+            protected void setUid(float uid) {
                 this.uid = uid;
             }
